system5-contract-auto/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
jinja2==3.1.4
python-dotenv==1.0.1
httpx==0.27.0
